Refactor: extract message and rank logic into service classes
