+++ src/main/java/com/xiluiis/service/MessageService.java
@@ -1,0 +1,5 @@
+package com.xiluiis.service;
+
+public interface MessageService {
+    String createMessage(String playerNameString, String pathString);
+}
