+++ src/main/java/com/xiluiis/service/RankService.java
@@ -1,0 +1,6 @@
+package com.xiluiis.service;
+import org.bukkit.entity.Player;
+
+public interface RankService {
+    String getPlayerRank(Player player);
+}
